rename model files as the classes
